requirements.txt
annotated-types==0.7.0
asgiref==3.5.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.0.4
Faker==13.12.0
google-ai-generativelanguage==0.6.5
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
